+++ projects/mk-kit/src/public-api.ts
@@ -9,6 +9,7 @@
  */
 export * from '@mk-kit/ui/core';
 export * from '@mk-kit/ui/icon';
+export * from '@mk-kit/ui/icon/extended';
 export * from '@mk-kit/ui/chip';
 export * from '@mk-kit/ui/checkbox';
 export * from '@mk-kit/ui/button';